report-system-backend-development/report-system-backend-development/requirements.txt
# FastAPI Backend Dependencies

# ============================================================
# Core Framework
# ============================================================
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# ============================================================
# Database
# ============================================================
pymongo==4.9.1  # MongoDB driver
motor==3.6.0    # Async MongoDB driver

# ============================================================
# Data Validation
# ============================================================
pydantic==2.10.3
pydantic-settings==2.6.1

# ============================================================
# Document Generation
# ============================================================
python-pptx==0.6.23    # PPTX generation
comtypes==1.4.8        # PDF conversion (Windows - PowerPoint COM automation)

# ============================================================
# Image Processing
# ============================================================
Pillow==10.4.0

# ============================================================
# GIS & Map Generation
# ============================================================
geopandas==1.0.1
matplotlib==3.9.4
shapely==2.0.7
contextily==1.6.2
pyproj==3.6.1
adjustText==0.8        # For map label adjustments

# ============================================================
# Utilities
# ============================================================
python-dotenv==1.0.1
aiofiles==24.1.0

# ============================================================
# Development & Testing
# ============================================================
pytest==8.3.4
httpx==0.28.1